+++ apps/rich-text-versioning/test/locations/ConfigScreen.spec.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import { describe, expect, it, vi } from 'vitest';
 import { mockCma, mockSdk } from '../../test/mocks';
-import ConfigScreen from './ConfigScreen';
+import ConfigScreen from '../../src/locations/ConfigScreen';
 
 vi.mock('@contentful/react-apps-toolkit', () => ({
   useSDK: () => mockSdk,
